test: split watch and election tests to separate files
